backend/requirements.txt
Flask
flask_cors
openai
google-cloud-speech
google-cloud-texttospeech
